+++ test/src/dart_comments_test.dart
@@ -20,13 +20,13 @@ main() {
 
     test('escape-comment', () {
       var comment = new Comment('/* foobar */');
-      expect(comment.asDartDoc(0), equals('/**  / *  foobar  * /  */\n'));
+      expect(comment.asDartDoc(0), equals('///  / *  foobar  * /\n'));
     });
 
     test('one-line-comment', () {
       expectABC(Comment comment) {
-        expect(comment.asDartDoc(0), equals('/** ABC */\n'));
-        expect(comment.asDartDoc(2), equals('  /** ABC */\n'));
+        expect(comment.asDartDoc(0), equals('/// ABC\n'));
+        expect(comment.asDartDoc(2), equals('  /// ABC\n'));
       }
 
       expectABC(new Comment('ABC'));
@@ -37,10 +37,8 @@ main() {
     test('multi-line-comment', () {
       expectABCdef(Comment comment) {
         expect(comment.asDartDoc(0), equals('''
-/**
- * ABC
- * def
- */
+/// ABC
+/// def
 '''));
       }
 
@@ -51,17 +49,15 @@ main() {
 
     test('break-lines', () {
       var chars = ('A ' * ((80 - 7) ~/ 2)).trimRight();
-      var charsShortened = chars.substring(0, chars.length - 2);
+      var charsShortened = chars.substring(0, chars.length - 4);
       var comment = new Comment(chars);
 
       // [chars] fit on one line with indentation=0.
-      expect(comment.asDartDoc(0), equals('/** $chars */\n'));
+      expect(comment.asDartDoc(0), equals('/// $chars\n'));
 
       // Adding an indentation of 2 characters should make it a block comment.
       expect(comment.asDartDoc(2), equals('''
-  /**
-   * $chars
-   */
+  /// $chars
 '''));
 
       comment = new Comment('$chars\n\n$chars');
@@ -70,13 +66,11 @@ main() {
       // which has multiple lines.
       // Multiple independend lines should be treated equally.
       expect(comment.asDartDoc(8), equals('''
-        /**
-         * $charsShortened
-         * A
-         *
-         * $charsShortened
-         * A
-         */
+        /// $charsShortened
+        /// A A
+        ///
+        /// $charsShortened
+        /// A A
 '''));
     });
   });